learnpress: remove wrongly identified TODO
At first I thought the `course-enrolled` action was triggered in cases
where the user simply connected back through the enrollment form for a
course they were already enrolled with. It is technically true, but that
triggers a real new enrollment in learnpress that resets the user's
progress in that course, so it should be loged as such.